requirements.txt
lmdb~=1.3.0
numpy~=1.21.2
biopython